fix: fix the conditon check variable
This condition checks the length of 'param2' (password) instead of 'param3' (ent_username) before adding '-u' and 'ent_username' to the arguments. It should check if 'param3' is not empty.

This condition checks the length of 'param3' (ent_username) instead of 'param4' (ent_identity) before adding '-a' and 'ent_identity' to the arguments. It should check if 'param4' is not empty.